Circle fixes: sources-tree utils
