+++ src/devtools/client/debugger/src/utils/sources-tree/addToTree.js
@@ -4,10 +4,11 @@
 
 //
 
+import { isInvalidUrl } from "../source";
+
 import {
   nodeHasChildren,
   isPathDirectory,
-  isInvalidUrl,
   partIsFile,
   createSourceNode,
   createDirectoryNode,
