+++ src/devtools/client/debugger/src/utils/sources-tree/utils.js
@@ -8,7 +8,6 @@ import { parse } from "../../utils/url";
 
 import { isPretty } from "../source";
 import { getURL } from "./getURL";
-const IGNORED_URLS = ["debugger eval code", "XStringBundle"];
 
 export function nodeHasChildren(item) {
   return item.type == "directory" && Array.isArray(item.contents);
@@ -68,30 +67,6 @@ export function getSourceFromNode(item) {
 
 export function isSource(item) {
   return item.type === "source";
-}
-
-export function getFileExtension(source) {
-  const { path } = getURL(source);
-  if (!path) {
-    return "";
-  }
-
-  const lastIndex = path.lastIndexOf(".");
-  return lastIndex !== -1 ? path.slice(lastIndex + 1) : "";
-}
-
-export function isNotJavaScript(source) {
-  return ["css", "svg", "png"].includes(getFileExtension(source));
-}
-
-export function isInvalidUrl(url, source) {
-  return (
-    !source.url ||
-    !url.group ||
-    isNotJavaScript(source) ||
-    IGNORED_URLS.includes(url) ||
-    isPretty(source)
-  );
 }
 
 export function partIsFile(index, parts, url) {
